+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformPlugin.kt
@@ -34,7 +34,6 @@ abstract class IntelliJPlatformPlugin : Plugin<Project> {
 
             // Test
             TestIdePerformanceTask,
-            TestIdeUiTask,
 
             // Run
             RunIdeTask,